+++ packages/network/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pact-toolbox/network
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @pact-toolbox/prelude@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes