+++ packages/unplugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pact-toolbox/unplugin
 
+## 0.0.8
+
+### Patch Changes
+
+- @pact-toolbox/network@0.0.7
+
 ## 0.0.7
 
 ### Patch Changes